server/requirements.txt
openai~=1.59.7
python-dotenv~=1.0.1
yfinance~=0.2.51
requests~=2.32.3
beautifulsoup4~=4.12.3
pandas~=2.2.3
flask
flask-cors